Module aws_sdk_kinesisanalyticsv2::error

Modules

code_validation_exception

See CodeValidationException

concurrent_modification_exception

See ConcurrentModificationException

invalid_application_configuration_exception

See InvalidApplicationConfigurationException

invalid_argument_exception

See InvalidArgumentException

invalid_request_exception

See InvalidRequestException

limit_exceeded_exception

See LimitExceededException

resource_in_use_exception

See ResourceInUseException

resource_not_found_exception

See ResourceNotFoundException

resource_provisioned_throughput_exceeded_exception

See ResourceProvisionedThroughputExceededException

service_unavailable_exception

See ServiceUnavailableException

too_many_tags_exception

See TooManyTagsException

unable_to_detect_schema_exception

See UnableToDetectSchemaException

unsupported_operation_exception

See UnsupportedOperationException

Structs

AddApplicationCloudWatchLoggingOptionError

AddApplicationInputError

AddApplicationInputProcessingConfigurationError

AddApplicationOutputError

AddApplicationReferenceDataSourceError

AddApplicationVpcConfigurationError

CodeValidationException

The user-provided application code (query) is not valid. This can be a simple syntax

ConcurrentModificationException

Exception thrown as a result of concurrent modifications to an application. This error can

CreateApplicationError

CreateApplicationPresignedUrlError

CreateApplicationSnapshotError

DeleteApplicationCloudWatchLoggingOptionError

DeleteApplicationError

DeleteApplicationInputProcessingConfigurationError

DeleteApplicationOutputError

DeleteApplicationReferenceDataSourceError

DeleteApplicationSnapshotError

DeleteApplicationVpcConfigurationError

DescribeApplicationError

DescribeApplicationSnapshotError

DescribeApplicationVersionError

DiscoverInputSchemaError

InvalidApplicationConfigurationException

The user-provided application configuration is not valid.

InvalidArgumentException

The specified input parameter value is not valid.

InvalidRequestException

The request JSON is not valid for the operation.

LimitExceededException

The number of allowed resources has been exceeded.

ListApplicationSnapshotsError

ListApplicationVersionsError

ListApplicationsError

ListTagsForResourceError

ResourceInUseException

The application is not available for this operation.

ResourceNotFoundException

Specified application can't be found.

ResourceProvisionedThroughputExceededException

Discovery failed to get a record from the streaming source because of the Kinesis

RollbackApplicationError

ServiceUnavailableException

The service cannot complete the request.

StartApplicationError

StopApplicationError

TagResourceError

TooManyTagsException

Application created with too many tags, or too many tags added to an application. Note that the maximum

UnableToDetectSchemaException

The data format is not valid. Kinesis Data Analytics cannot detect the schema for

UnsupportedOperationException

The request was rejected because a specified parameter is not supported or a specified resource is not valid for this

UntagResourceError

UpdateApplicationError

UpdateApplicationMaintenanceConfigurationError

Enums

AddApplicationCloudWatchLoggingOptionErrorKind

AddApplicationInputErrorKind

AddApplicationInputProcessingConfigurationErrorKind

AddApplicationOutputErrorKind

AddApplicationReferenceDataSourceErrorKind

AddApplicationVpcConfigurationErrorKind

CreateApplicationErrorKind

CreateApplicationPresignedUrlErrorKind

CreateApplicationSnapshotErrorKind

DeleteApplicationCloudWatchLoggingOptionErrorKind

DeleteApplicationErrorKind

DeleteApplicationInputProcessingConfigurationErrorKind

DeleteApplicationOutputErrorKind

DeleteApplicationReferenceDataSourceErrorKind

DeleteApplicationSnapshotErrorKind

DeleteApplicationVpcConfigurationErrorKind

DescribeApplicationErrorKind

DescribeApplicationSnapshotErrorKind

DescribeApplicationVersionErrorKind

DiscoverInputSchemaErrorKind

ListApplicationSnapshotsErrorKind

ListApplicationVersionsErrorKind

ListApplicationsErrorKind

ListTagsForResourceErrorKind

RollbackApplicationErrorKind

StartApplicationErrorKind

StopApplicationErrorKind

TagResourceErrorKind

UntagResourceErrorKind

UpdateApplicationErrorKind

UpdateApplicationMaintenanceConfigurationErrorKind