Module aws_sdk_kinesisanalyticsv2::model

Modules

application_code_configuration

See ApplicationCodeConfiguration

application_code_configuration_description

See ApplicationCodeConfigurationDescription

application_code_configuration_update

See ApplicationCodeConfigurationUpdate

application_configuration

See ApplicationConfiguration

application_configuration_description

See ApplicationConfigurationDescription

application_configuration_update

See ApplicationConfigurationUpdate

application_detail

See ApplicationDetail

application_maintenance_configuration_description

See ApplicationMaintenanceConfigurationDescription

application_maintenance_configuration_update

See ApplicationMaintenanceConfigurationUpdate

application_restore_configuration

See ApplicationRestoreConfiguration

application_snapshot_configuration

See ApplicationSnapshotConfiguration

application_snapshot_configuration_description

See ApplicationSnapshotConfigurationDescription

application_snapshot_configuration_update

See ApplicationSnapshotConfigurationUpdate

application_summary

See ApplicationSummary

application_version_summary

See ApplicationVersionSummary

catalog_configuration

See CatalogConfiguration

catalog_configuration_description

See CatalogConfigurationDescription

catalog_configuration_update

See CatalogConfigurationUpdate

checkpoint_configuration

See CheckpointConfiguration

checkpoint_configuration_description

See CheckpointConfigurationDescription

checkpoint_configuration_update

See CheckpointConfigurationUpdate

cloud_watch_logging_option

See CloudWatchLoggingOption

cloud_watch_logging_option_description

See CloudWatchLoggingOptionDescription

cloud_watch_logging_option_update

See CloudWatchLoggingOptionUpdate

code_content

See CodeContent

code_content_description

See CodeContentDescription

code_content_update

See CodeContentUpdate

csv_mapping_parameters

See CsvMappingParameters

custom_artifact_configuration

See CustomArtifactConfiguration

custom_artifact_configuration_description

See CustomArtifactConfigurationDescription

deploy_as_application_configuration

See DeployAsApplicationConfiguration

deploy_as_application_configuration_description

See DeployAsApplicationConfigurationDescription

deploy_as_application_configuration_update

See DeployAsApplicationConfigurationUpdate

destination_schema

See DestinationSchema

environment_properties

See EnvironmentProperties

environment_property_descriptions

See EnvironmentPropertyDescriptions

environment_property_updates

See EnvironmentPropertyUpdates

flink_application_configuration

See FlinkApplicationConfiguration

flink_application_configuration_description

See FlinkApplicationConfigurationDescription

flink_application_configuration_update

See FlinkApplicationConfigurationUpdate

flink_run_configuration

See FlinkRunConfiguration

glue_data_catalog_configuration

See GlueDataCatalogConfiguration

glue_data_catalog_configuration_description

See GlueDataCatalogConfigurationDescription

glue_data_catalog_configuration_update

See GlueDataCatalogConfigurationUpdate

input

See Input

input_description

See InputDescription

input_lambda_processor

See InputLambdaProcessor

input_lambda_processor_description

See InputLambdaProcessorDescription

input_lambda_processor_update

See InputLambdaProcessorUpdate

input_parallelism

See InputParallelism

input_parallelism_update

See InputParallelismUpdate

input_processing_configuration

See InputProcessingConfiguration

input_processing_configuration_description

See InputProcessingConfigurationDescription

input_processing_configuration_update

See InputProcessingConfigurationUpdate

input_schema_update

See InputSchemaUpdate

input_starting_position_configuration

See InputStartingPositionConfiguration

input_update

See InputUpdate

json_mapping_parameters

See JsonMappingParameters

kinesis_firehose_input

See KinesisFirehoseInput

kinesis_firehose_input_description

See KinesisFirehoseInputDescription

kinesis_firehose_input_update

See KinesisFirehoseInputUpdate

kinesis_firehose_output

See KinesisFirehoseOutput

kinesis_firehose_output_description

See KinesisFirehoseOutputDescription

kinesis_firehose_output_update

See KinesisFirehoseOutputUpdate

kinesis_streams_input

See KinesisStreamsInput

kinesis_streams_input_description

See KinesisStreamsInputDescription

kinesis_streams_input_update

See KinesisStreamsInputUpdate

kinesis_streams_output

See KinesisStreamsOutput

kinesis_streams_output_description

See KinesisStreamsOutputDescription

kinesis_streams_output_update

See KinesisStreamsOutputUpdate

lambda_output

See LambdaOutput

lambda_output_description

See LambdaOutputDescription

lambda_output_update

See LambdaOutputUpdate

mapping_parameters

See MappingParameters

maven_reference

See MavenReference

monitoring_configuration

See MonitoringConfiguration

monitoring_configuration_description

See MonitoringConfigurationDescription

monitoring_configuration_update

See MonitoringConfigurationUpdate

output

See Output

output_description

See OutputDescription

output_update

See OutputUpdate

parallelism_configuration

See ParallelismConfiguration

parallelism_configuration_description

See ParallelismConfigurationDescription

parallelism_configuration_update

See ParallelismConfigurationUpdate

property_group

See PropertyGroup

record_column

See RecordColumn

record_format

See RecordFormat

reference_data_source

See ReferenceDataSource

reference_data_source_description

See ReferenceDataSourceDescription

reference_data_source_update

See ReferenceDataSourceUpdate

run_configuration

See RunConfiguration

run_configuration_description

See RunConfigurationDescription

run_configuration_update

See RunConfigurationUpdate

s3_application_code_location_description

See S3ApplicationCodeLocationDescription

s3_configuration

See S3Configuration

s3_content_base_location

See S3ContentBaseLocation

s3_content_base_location_description

See S3ContentBaseLocationDescription

s3_content_base_location_update

See S3ContentBaseLocationUpdate

s3_content_location

See S3ContentLocation

s3_content_location_update

See S3ContentLocationUpdate

s3_reference_data_source

See S3ReferenceDataSource

s3_reference_data_source_description

See S3ReferenceDataSourceDescription

s3_reference_data_source_update

See S3ReferenceDataSourceUpdate

snapshot_details

See SnapshotDetails

source_schema

See SourceSchema

sql_application_configuration

See SqlApplicationConfiguration

sql_application_configuration_description

See SqlApplicationConfigurationDescription

sql_application_configuration_update

See SqlApplicationConfigurationUpdate

sql_run_configuration

See SqlRunConfiguration

tag

See Tag

vpc_configuration

See VpcConfiguration

vpc_configuration_description

See VpcConfigurationDescription

vpc_configuration_update

See VpcConfigurationUpdate

zeppelin_application_configuration

See ZeppelinApplicationConfiguration

zeppelin_application_configuration_description

See ZeppelinApplicationConfigurationDescription

zeppelin_application_configuration_update

See ZeppelinApplicationConfigurationUpdate

zeppelin_monitoring_configuration

See ZeppelinMonitoringConfiguration

zeppelin_monitoring_configuration_description

See ZeppelinMonitoringConfigurationDescription

zeppelin_monitoring_configuration_update

See ZeppelinMonitoringConfigurationUpdate

Structs

ApplicationCodeConfiguration

Describes code configuration for an application.

ApplicationCodeConfigurationDescription

Describes code configuration for an application.

ApplicationCodeConfigurationUpdate

Describes code configuration updates for an application. This is supported for a Flink-based Kinesis Data Analytics application or a SQL-based Kinesis Data Analytics application.

ApplicationConfiguration

Specifies the creation parameters for a Kinesis Data Analytics application.

ApplicationConfigurationDescription

Describes details about the application code and starting parameters for a Kinesis Data Analytics application.

ApplicationConfigurationUpdate

Describes updates to an application's configuration.

ApplicationDetail

Describes the application, including the application Amazon Resource Name (ARN), status,

ApplicationMaintenanceConfigurationDescription

The details of the maintenance configuration for the application.

ApplicationMaintenanceConfigurationUpdate

Describes the updated maintenance configuration for the application.

ApplicationRestoreConfiguration

Specifies the method and snapshot to use when restarting an application using previously saved application state.

ApplicationSnapshotConfiguration

Describes whether snapshots are enabled for a Flink-based Kinesis Data Analytics application.

ApplicationSnapshotConfigurationDescription

Describes whether snapshots are enabled for a Flink-based Kinesis Data Analytics application.

ApplicationSnapshotConfigurationUpdate

Describes updates to whether snapshots are enabled for a Flink-based Kinesis Data Analytics application.

ApplicationSummary

Provides application summary information, including the application Amazon Resource Name (ARN), name, and status.

ApplicationVersionSummary

The summary of the application version.

CatalogConfiguration

The configuration parameters for the default AWS Glue database. You use this database for SQL queries that you write in a Kinesis Data Analytics Studio notebook.

CatalogConfigurationDescription

The configuration parameters for the default AWS Glue database. You use this database for Apache Flink SQL queries and table API transforms that you write in a Kinesis Data Analytics Studio notebook.

CatalogConfigurationUpdate

Updates to

CheckpointConfiguration

Describes an application's checkpointing configuration. Checkpointing is the process of persisting application state for fault

CheckpointConfigurationDescription

Describes checkpointing parameters for a Flink-based Kinesis Data Analytics application.

CheckpointConfigurationUpdate

Describes updates to the checkpointing parameters for a Flink-based Kinesis Data Analytics application.

CloudWatchLoggingOption

Provides a description of Amazon CloudWatch logging options, including the log stream

CloudWatchLoggingOptionDescription

Describes the Amazon CloudWatch logging option.

CloudWatchLoggingOptionUpdate

Describes the Amazon CloudWatch logging option updates.

CodeContent

Specifies either the application code, or the location of the application code, for a

CodeContentDescription

Describes details about the code of a Kinesis Data Analytics application.

CodeContentUpdate

Describes an update to the code of an application. Not supported for Apache Zeppelin.

CsvMappingParameters

For a SQL-based Kinesis Data Analytics application, provides additional mapping information when the record

CustomArtifactConfiguration

Specifies dependency JARs, as well as JAR files that contain user-defined functions (UDF).

CustomArtifactConfigurationDescription

Specifies a dependency JAR or a JAR of user-defined functions.

DeployAsApplicationConfiguration

The information required to deploy a Kinesis Data Analytics Studio notebook as an application with durable state..

DeployAsApplicationConfigurationDescription

The configuration information required to deploy an Amazon Data Analytics Studio notebook as an application with durable state.

DeployAsApplicationConfigurationUpdate

Updates to the configuration information required to deploy an Amazon Data Analytics Studio notebook as an application with durable state..

DestinationSchema

Describes the data format when records are written to the destination in a SQL-based Kinesis Data Analytics application.

EnvironmentProperties

Describes execution properties for a Flink-based Kinesis Data Analytics application.

EnvironmentPropertyDescriptions

Describes the execution properties for an Apache Flink runtime.

EnvironmentPropertyUpdates

Describes updates to the execution property groups for a Flink-based Kinesis Data Analytics application or a Studio notebook.

FlinkApplicationConfiguration

Describes configuration parameters for a Flink-based Kinesis Data Analytics application or a Studio notebook.

FlinkApplicationConfigurationDescription

Describes configuration parameters for a Flink-based Kinesis Data Analytics application.

FlinkApplicationConfigurationUpdate

Describes updates to the configuration parameters for a Flink-based Kinesis Data Analytics application.

FlinkRunConfiguration

Describes the starting parameters for a Flink-based Kinesis Data Analytics application.

GlueDataCatalogConfiguration

The configuration of the Glue Data Catalog that you use for Apache Flink SQL queries and table API transforms that you write in an application.

GlueDataCatalogConfigurationDescription

The configuration of the Glue Data Catalog that you use for Apache Flink SQL queries and table API transforms that you write in an application.

GlueDataCatalogConfigurationUpdate

Updates to the configuration of the Glue Data Catalog that you use for SQL queries that you write in a Kinesis Data Analytics Studio notebook.

Input

When you configure the application input for a SQL-based Kinesis Data Analytics application, you specify the streaming source, the in-application stream

InputDescription

Describes the application input configuration for a SQL-based Kinesis Data Analytics application.

InputLambdaProcessor

An object that contains the Amazon Resource Name (ARN) of the AWS Lambda function that is

InputLambdaProcessorDescription

For a SQL-based Kinesis Data Analytics application, an object that contains the

InputLambdaProcessorUpdate

For a SQL-based Kinesis Data Analytics application, represents an update to the

InputParallelism

For a SQL-based Kinesis Data Analytics application, describes the number of

InputParallelismUpdate

For a SQL-based Kinesis Data Analytics application, provides updates to the

InputProcessingConfiguration

For a SQL-based Kinesis Data Analytics application, describes a processor that is

InputProcessingConfigurationDescription

For a SQL-based Kinesis Data Analytics application, provides the configuration

InputProcessingConfigurationUpdate

For a SQL-based Kinesis Data Analytics application, describes updates to an InputProcessingConfiguration.

InputSchemaUpdate

Describes updates for an SQL-based Kinesis Data Analytics application's input

InputStartingPositionConfiguration

Describes the point at which the application reads from

InputUpdate

For a SQL-based Kinesis Data Analytics application, describes updates to a

JsonMappingParameters

For a SQL-based Kinesis Data Analytics application, provides additional mapping

KinesisFirehoseInput

For a SQL-based Kinesis Data Analytics application, identifies a Kinesis Data

KinesisFirehoseInputDescription

Describes the Amazon Kinesis Data Firehose delivery stream that is configured as the streaming source

KinesisFirehoseInputUpdate

For a SQL-based Kinesis Data Analytics application, when updating application

KinesisFirehoseOutput

For a SQL-based Kinesis Data Analytics application, when configuring application

KinesisFirehoseOutputDescription

For a SQL-based Kinesis Data Analytics application's output, describes the Kinesis

KinesisFirehoseOutputUpdate

For a SQL-based Kinesis Data Analytics application, when updating an output

KinesisStreamsInput

Identifies a Kinesis data stream as the streaming source. You provide the

KinesisStreamsInputDescription

For a SQL-based Kinesis Data Analytics application, describes the Kinesis data

KinesisStreamsInputUpdate

When you update the input configuration for a SQL-based Kinesis Data Analytics application, provides information about a

KinesisStreamsOutput

When you configure a SQL-based Kinesis Data Analytics application's output,

KinesisStreamsOutputDescription

For an SQL-based Kinesis Data Analytics application's output, describes the Kinesis

KinesisStreamsOutputUpdate

When you update a SQL-based Kinesis Data Analytics application's output

LambdaOutput

When you configure a SQL-based Kinesis Data Analytics application's output,

LambdaOutputDescription

For a SQL-based Kinesis Data Analytics application's output, describes the AWS

LambdaOutputUpdate

When you update an SQL-based Kinesis Data Analytics application's output

MappingParameters

When you configure a SQL-based Kinesis Data Analytics application's input at the

MavenReference

The information required to specify a Maven reference. You can use Maven references to specify dependency JAR files.

MonitoringConfiguration

Describes configuration parameters for Amazon CloudWatch logging for an application. For more information about CloudWatch logging, see Monitoring.

MonitoringConfigurationDescription

Describes configuration parameters for CloudWatch logging for an application.

MonitoringConfigurationUpdate

Describes updates to configuration parameters for Amazon CloudWatch logging for an application.

Output

Describes a SQL-based Kinesis Data Analytics application's output configuration,

OutputDescription

For a SQL-based Kinesis Data Analytics application, describes the application

OutputUpdate

For a SQL-based Kinesis Data Analytics application, describes updates to the

ParallelismConfiguration

Describes parameters for how a Flink-based Kinesis Data Analytics application

ParallelismConfigurationDescription

Describes parameters for how a Flink-based Kinesis Data Analytics application

ParallelismConfigurationUpdate

Describes updates to parameters for how an application executes multiple tasks simultaneously.

PropertyGroup

Property key-value pairs passed into an application.

RecordColumn

For a SQL-based Kinesis Data Analytics application, describes the mapping of each

RecordFormat

For a SQL-based Kinesis Data Analytics application, describes the record format

ReferenceDataSource

For a SQL-based Kinesis Data Analytics application, describes the reference data

ReferenceDataSourceDescription

For a SQL-based Kinesis Data Analytics application, describes the reference data

ReferenceDataSourceUpdate

When you update a reference data source configuration for a

RunConfiguration

Describes the starting parameters for an Kinesis Data Analytics application.

RunConfigurationDescription

Describes the starting properties for a Kinesis Data Analytics application.

RunConfigurationUpdate

Describes the updates to the starting parameters for a Kinesis Data Analytics application.

S3ApplicationCodeLocationDescription

Describes the location of an application's code stored in an S3 bucket.

S3Configuration

For a SQL-based Kinesis Data Analytics application, provides a description of an

S3ContentBaseLocation

The S3 bucket that holds the application information.

S3ContentBaseLocationDescription

The description of the S3 base location that holds the application.

S3ContentBaseLocationUpdate

The information required to update the S3 base location that holds the application.

S3ContentLocation

For a Kinesis Data Analytics application provides a

S3ContentLocationUpdate

Describes an update for the Amazon S3 code content location for an application.

S3ReferenceDataSource

For a SQL-based Kinesis Data Analytics application, identifies the Amazon S3

S3ReferenceDataSourceDescription

For a SQL-based Kinesis Data Analytics application, provides the bucket name and

S3ReferenceDataSourceUpdate

For a SQL-based Kinesis Data Analytics application, describes the Amazon S3 bucket

SnapshotDetails

Provides details about a snapshot of application state.

SourceSchema

For a SQL-based Kinesis Data Analytics application, describes the format of the

SqlApplicationConfiguration

Describes the inputs, outputs, and reference data sources for a SQL-based Kinesis Data Analytics application.

SqlApplicationConfigurationDescription

Describes the inputs, outputs, and reference data sources for a SQL-based Kinesis Data Analytics application.

SqlApplicationConfigurationUpdate

Describes updates to the input streams, destination streams, and reference data sources

SqlRunConfiguration

Describes the starting parameters for a SQL-based Kinesis Data Analytics application.

Tag

A key-value pair (the value is optional) that you can define and assign to AWS resources.

VpcConfiguration

Describes the parameters of a VPC used by the application.

VpcConfigurationDescription

Describes the parameters of a VPC used by the application.

VpcConfigurationUpdate

Describes updates to the VPC configuration used by the application.

ZeppelinApplicationConfiguration

The configuration of a Kinesis Data Analytics Studio notebook.

ZeppelinApplicationConfigurationDescription

The configuration of a Kinesis Data Analytics Studio notebook.

ZeppelinApplicationConfigurationUpdate

Updates to the configuration of Kinesis Data Analytics Studio notebook.

ZeppelinMonitoringConfiguration

Describes configuration parameters for Amazon CloudWatch logging for a Kinesis Data Analytics Studio notebook. For more information about CloudWatch logging, see Monitoring.

ZeppelinMonitoringConfigurationDescription

The monitoring configuration for Apache Zeppelin within a Kinesis Data Analytics Studio notebook.

ZeppelinMonitoringConfigurationUpdate

Updates to the monitoring configuration for Apache Zeppelin within a Kinesis Data Analytics Studio notebook.

Enums

ApplicationMode

ApplicationRestoreType

ApplicationStatus

ArtifactType

CodeContentType

ConfigurationType

InputStartingPosition

LogLevel

MetricsLevel

RecordFormatType

RuntimeEnvironment

SnapshotStatus

UrlType