Module aws_sdk_kinesisanalyticsv2::operation

Structs

AddApplicationCloudWatchLoggingOption

Adds an Amazon CloudWatch log stream to monitor application configuration errors.

AddApplicationInput

AddApplicationInputProcessingConfiguration

Adds an InputProcessingConfiguration to a SQL-based Kinesis Data Analytics application. An input processor pre-processes records

AddApplicationOutput

Adds an external destination to your SQL-based Kinesis Data Analytics application.

AddApplicationReferenceDataSource

Adds a reference data source to an existing SQL-based Kinesis Data Analytics application.

AddApplicationVpcConfiguration

Adds a Virtual Private Cloud (VPC) configuration to the application. Applications can use VPCs to store

CreateApplication

Creates a Kinesis Data Analytics application. For information about creating a

CreateApplicationPresignedUrl

Creates and returns a URL that you can use to connect to

CreateApplicationSnapshot

Creates a snapshot of the application's state data.

DeleteApplication

Deletes the specified application. Kinesis Data Analytics halts application execution and deletes the application.

DeleteApplicationCloudWatchLoggingOption

Deletes an Amazon CloudWatch log stream from an Kinesis Data Analytics application.

DeleteApplicationInputProcessingConfiguration

Deletes an InputProcessingConfiguration from an input.

DeleteApplicationOutput

Deletes the output destination configuration from your SQL-based Kinesis Data Analytics application's configuration.

DeleteApplicationReferenceDataSource

Deletes a reference data source configuration from the specified SQL-based Kinesis Data Analytics application's configuration.

DeleteApplicationSnapshot

Deletes a snapshot of application state.

DeleteApplicationVpcConfiguration

Removes a VPC configuration from a Kinesis Data Analytics application.

DescribeApplication

Returns information about a specific Kinesis Data Analytics application.

DescribeApplicationSnapshot

Returns information about a snapshot of application state data.

DescribeApplicationVersion

Provides a detailed description of a specified version of the application. To see a list of all the versions of an application, invoke the ListApplicationVersions operation.

DiscoverInputSchema

Infers a schema for a SQL-based Kinesis Data Analytics application by evaluating

ListApplicationSnapshots

Lists information about the current application snapshots.

ListApplicationVersions

Lists all the versions for the specified application, including versions that were rolled back. The response also includes a summary of the configuration

ListApplications

Returns a list of Kinesis Data Analytics applications in your account. For each

ListTagsForResource

Retrieves the list of key-value tags assigned to the application. For more information, see

RollbackApplication

Reverts the application to the previous running version. You can

StartApplication

Starts the specified Kinesis Data Analytics application. After creating an application, you must exclusively call this operation to

StopApplication

Stops the application from processing data. You can stop

TagResource

Adds one or more key-value tags to a Kinesis Data Analytics application. Note that the maximum number of application

UntagResource

Removes one or more tags from a Kinesis Data Analytics application. For more information, see

UpdateApplication

Updates an existing Kinesis Data Analytics application. Using this operation, you

UpdateApplicationMaintenanceConfiguration

Updates the maintenance configuration of the Kinesis Data Analytics application.